% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a spatial model with Vecchia approximation}
\usage{
fit_model(
  Y_obs,
  locs_obs,
  X_obs = NULL,
  svc_indices = NULL,
  censored_indices,
  M,
  bayesian = FALSE,
  chains = NULL,
  iter_warmup = NULL,
  iter_sampling = NULL,
  parallel_chains = NULL,
  fixed_range = TRUE
)
}
\arguments{
\item{Y_obs}{Observed response vector}

\item{locs_obs}{Matrix of spatial locations}

\item{X_obs}{Design matrix (optional)}

\item{svc_indices}{Indices for spatially varying coefficients}

\item{censored_indices}{Binary vector indicating censored observations}

\item{M}{Number of nearest neighbors}

\item{bayesian}{Logical, use Bayesian estimation}

\item{chains}{Number of MCMC chains}

\item{iter_warmup}{Warmup iterations}

\item{iter_sampling}{Sampling iterations}

\item{parallel_chains}{Number of parallel chains}

\item{fixed_range}{Use fixed range parameter}
}
\value{
A fitted model object
}
\description{
Fit a spatial model with Vecchia approximation
}
