% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{prediction}
\alias{prediction}
\title{Make predictions from a fitted spatial model}
\usage{
prediction(
  fit_object,
  locs_pred,
  X_pred,
  M,
  simulations = FALSE,
  n_simulations = NULL,
  pred_coef = FALSE
)
}
\arguments{
\item{fit_object}{A fitted model object from fit_model()}

\item{locs_pred}{Matrix of prediction locations}

\item{X_pred}{Design matrix for prediction locations}

\item{M}{Number of nearest neighbors for Vecchia approximation}

\item{simulations}{Logical, generate conditional simulations}

\item{n_simulations}{Number of simulations (for non-Bayesian models)}

\item{pred_coef}{Logical, predict spatially varying coefficients (SVC only)}
}
\value{
A list containing predictions and optionally simulations/coefficients
}
\description{
Make predictions from a fitted spatial model
}
