% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_SVC}
\alias{test_SVC}
\title{Test dataset for Spatially Varying Coefficients model}
\format{
A data frame with 100 rows and 7 columns:
\describe{
  \item{X}{X spatial coordinate (1 to 100)}
  \item{Y}{Y spatial coordinate (1 to 100)}
  \item{X1}{Transformed covariate: sqrt(|Y - 60|)}
  \item{X2}{Transformed covariate: sqrt(|X - 20|)}
  \item{Z}{True (uncensored) response variable}
  \item{censored}{Censored response variable}
  \item{is_censored}{Binary indicator: 1 if censored, 0 otherwise}
}
}
\source{
Simulated data for package demonstration
}
\usage{
test_SVC
}
\description{
Hold-out test set corresponding to \code{\link{grille_SVC}}, containing
20% of the original simulated data for model validation.
}
\details{
This is the test split from the same simulation that produced \code{\link{grille_SVC}}.
Use this dataset to evaluate SVC model prediction performance.
}
\examples{
# Use with grille_SVC for prediction
data(grille_SVC)
data(test_SVC)

# Fit SVC model on training data
fit_svc <- fit_model(
  Y_obs = grille_SVC$censored,
  locs_obs = as.matrix(grille_SVC[, c("X", "Y")]),
  X_obs = as.matrix(grille_SVC[, c("X1", "X2")]),
  svc_indices = c(1, 2),
  censored_indices = grille_SVC$is_censored,
  M = 30,
  bayesian = FALSE
)

# Make predictions on test set (when predict function is available)
# predictions <- predict(fit_svc,
#                        newlocs = as.matrix(test_SVC[, c("X", "Y")]),
#                        newX = as.matrix(test_SVC[, c("X1", "X2")]))

}
\seealso{
\code{\link{grille_SVC}}
}
\keyword{datasets}
