% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{package}
\name{VecchiaCensored}
\alias{VecchiaCensored-package}
\alias{VecchiaCensored}
\title{VecchiaCensored: Vecchia Approximation for Censored Spatial Data}
\description{
This package implements Vecchia approximation methods for spatial
statistics with censored data. It provides efficient algorithms using
sparse matrix decompositions and Rcpp for computational performance.
}
\author{
\strong{Maintainer}: Yacine IDIR \email{email@example.com}

}
